requirements.txt
Pillow
numpy
pytest
pylint
tensorboard_logger
tensorboard-pytorch
tensorboard
future